Refactor StarkNetBitcoinDepositor and Mock Contracts
- Cleaned up whitespace and formatting across multiple files for consistency.
- Enhanced error handling in MockTBTCBridge and MockStarkGateBridge for better testing.
- Introduced MockTBTCBridgeWithSweep to facilitate testing of deposit sweeps.
- Updated tests to ensure proper initialization and functionality of deposit-related features.
- Improved event emissions and assertions in tests for clarity and accuracy.

+++ solidity/test/cross-chain/starknet/FeeManagement.test.sol
@@ -14,24 +14,24 @@ contract FeeManagementTest is Test {
     MockTBTCVault public mockVault;
     MockTBTCToken public mockToken;
     MockStarkGateBridge public mockStarkGateBridge;
-    
+
     address public owner;
     address public nonOwner;
-    
+
     uint256 public constant INITIAL_FEE = 0.01 ether;
-    
+
     event L1ToL2MessageFeeUpdated(uint256 newFee);
-    
+
     function setUp() public {
         owner = address(this);
         nonOwner = address(0x1234);
-        
+
         // Deploy mocks
         mockToken = new MockTBTCToken();
         mockBridge = new MockTBTCBridge();
         mockVault = new MockTBTCVault(address(mockToken));
         mockStarkGateBridge = new MockStarkGateBridge();
-        
+
         // Deploy depositor
         depositor = new StarkNetBitcoinDepositor(
             address(mockBridge),
@@ -40,66 +40,66 @@ contract FeeManagementTest is Test {
             INITIAL_FEE
         );
     }
-    
+
     function testUpdateL1ToL2MessageFeeSuccess() public {
         uint256 newFee = 0.02 ether;
-        
+
         // Expect event emission
         vm.expectEmit(true, true, true, true);
         emit L1ToL2MessageFeeUpdated(newFee);
-        
+
         // Update fee
         depositor.updateL1ToL2MessageFee(newFee);
-        
+
         // Verify fee was updated
         assertEq(depositor.l1ToL2MessageFee(), newFee);
     }
-    
+
     function testUpdateL1ToL2MessageFeeAccessControl() public {
         uint256 newFee = 0.02 ether;
-        
+
         // Try to update as non-owner
         vm.prank(nonOwner);
         vm.expectRevert("Ownable: caller is not the owner");
         depositor.updateL1ToL2MessageFee(newFee);
-        
+
         // Verify fee was not changed
         assertEq(depositor.l1ToL2MessageFee(), INITIAL_FEE);
     }
-    
+
     function testUpdateL1ToL2MessageFeeZeroValue() public {
         // Try to set zero fee
         vm.expectRevert("Fee must be greater than 0");
         depositor.updateL1ToL2MessageFee(0);
-        
+
         // Verify fee was not changed
         assertEq(depositor.l1ToL2MessageFee(), INITIAL_FEE);
     }
-    
+
     function testQuoteFinalizeDeposit() public {
         // Test initial fee
         assertEq(depositor.quoteFinalizeDeposit(), INITIAL_FEE);
-        
+
         // Update fee and test again
         uint256 newFee = 0.03 ether;
         depositor.updateL1ToL2MessageFee(newFee);
         assertEq(depositor.quoteFinalizeDeposit(), newFee);
     }
-    
+
     function testMultipleFeeUpdates() public {
         uint256[] memory fees = new uint256[](3);
         fees[0] = 0.02 ether;
         fees[1] = 0.015 ether;
         fees[2] = 0.025 ether;
-        
-        for (uint i = 0; i < fees.length; i++) {
+
+        for (uint256 i = 0; i < fees.length; i++) {
             // Expect event
             vm.expectEmit(true, true, true, true);
             emit L1ToL2MessageFeeUpdated(fees[i]);
-            
+
             // Update fee
             depositor.updateL1ToL2MessageFee(fees[i]);
-            
+
             // Verify
             assertEq(depositor.l1ToL2MessageFee(), fees[i]);
         }
